requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
decorator==5.1.1
Django==4.2.16
django-allauth==65.2.0
django-celery-beat==2.7.0
django-filter==24.3
django-geojson==4.1.0
django-humanize==0.1.2
django-leaflet==0.31.0
django-simple-history==3.7.0
django-timezone-field==7.0
django-unicorn==0.61.0
djangorestframework==3.15.2
Faker==33.0.0
fastkml==0.12
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
humanize==4.11.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
kombu==5.4.2
Markdown==3.7
numpy==2.0.2
orjson==3.10.11
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pygeoif==0.7
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
rfc3986==1.5.0
shapely==2.0.6
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
zipp==3.21.0
